move pagination tests to their own file
